configure.ac: split lib/include fuse
osxfuse is rarely in the same top directory. The popular binary installer
for example installs the libs in "/usr/local/lib" but the headers in
"/usr/local/include/osxfuse/fuse".

Looking for the headers under the "--with-libfuse=" directory, presuming
that given directory is "/usr/local" most of the time, will fail.